ccimx8: use 64 bit variables when managing RAM sizes
Now that we have support for ccimx8x variants with 4 GiB of RAM (which requires
at least 33 bits to represent in bytes), we need to make sure that all
operations regarding RAM sizes in ccimx8 platforms use 64 bit variables so all
possible values can be represented.

https://onedigi.atlassian.net/browse/DEL-8604

+++ board/digi/common/hwid.h
@@ -41,7 +41,7 @@ int board_override_hwid(const struct digi_hwid *hwid);
 void board_update_hwid(bool is_fuse);
 int board_lock_hwid(void);
 void fdt_fixup_hwid(void *fdt, const struct digi_hwid *hwid);
-u32 hwid_get_ramsize(const struct digi_hwid *hwid);
+u64 hwid_get_ramsize(const struct digi_hwid *hwid);
 void print_hwid_hex(struct digi_hwid *hwid);
 
 #endif	/* __HWID_H_ */
